Refactor Spotify
