+++ music_assistant/providers/spotify/helpers.py
@@ -1,4 +1,4 @@
-"""Helpers/utils for the Spotify musicprovider."""
+"""NEW helpers.py Helpers/utils for the Spotify musicprovider."""
 
 from __future__ import annotations
 
@@ -23,10 +23,10 @@ async def get_librespot_binary() -> str:
     system = platform.system().lower().replace("darwin", "macos")
     architecture = platform.machine().lower()
 
-    if bridge_binary := await check_librespot(
+    if librespot_binary := await check_librespot(
         os.path.join(base_path, f"librespot-{system}-{architecture}")
     ):
-        return bridge_binary
+        return librespot_binary
 
     msg = f"Unable to locate Librespot for {system}/{architecture}"
     raise RuntimeError(msg)
